+++ src/pages/Comparisons/Comparison.js
@@ -97,6 +97,12 @@ function Comparison(props) {
         }
     }, [params.comparisonId, loadVersions]);
 
+    useEffect(() => {
+        if (metaData?.title) {
+            document.title = `${metaData.title} - Comparison - ORKG`;
+        }
+    }, [metaData]);
+
     /** adding some additional state for meta data **/
 
     const [cookies, setCookie] = useCookies();
@@ -414,7 +420,7 @@ function Comparison(props) {
                                         >
                                             Go back
                                         </span>{' '}
-                                        or <Link to={ROUTES.HOME}>go to the homepage {contributionsList.length}</Link>.
+                                        or <Link to={ROUTES.HOME}>go to the homepage</Link>.
                                     </>
                                 )}
                             </Alert>
@@ -525,7 +531,7 @@ function Comparison(props) {
                                     <li key={`ref${index}`}>
                                         <small>
                                             <i>
-                                                <ValuePlugins type="literal">{reference}</ValuePlugins>
+                                                <ValuePlugins type="literal">{reference.label}</ValuePlugins>
                                             </i>
                                         </small>
                                     </li>